feat(frontend): mysql、tendbcluster定点构造优化 #5611 # Reviewed, transaction id: 20639
