+++ dbm-ui/frontend/src/hooks/useTicketCloneInfo/generateCloneData/tendbcluster/rollback.ts
@@ -10,13 +10,13 @@
  * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
  * the specific language governing permissions and limitations under the License.
  */
-import type { MySQLRollbackDetails } from '@services/model/ticket/details/mysql';
+import type { SpiderRollbackDetails } from '@services/model/ticket/details/spider';
 import TicketModel from '@services/model/ticket/ticket';
 
 import { random } from '@utils';
 
 // Spider 定点构造
-export function generateSpiderRollbackCloneData(ticketData: TicketModel<MySQLRollbackDetails>) {
+export function generateSpiderRollbackCloneData(ticketData: TicketModel<SpiderRollbackDetails>) {
   const { clusters, infos, rollback_cluster_type } = ticketData.details;
   const tableDataList = infos.map((item) => ({
     rowKey: random(),
