+++ dbm-ui/frontend/src/locales/zh-cn.json
@@ -3418,6 +3418,8 @@
     "包含通配符时, 每一单元格只允许输入单个对象。% 不能独立使用， * 只能单独使用": "包含通配符时, 每一单元格只允许输入单个对象。% 不能独立使用， * 只能单独使用",
     "查看域名/IP对应关系": "查看域名/IP对应关系",
     "确认继续所有人工确认节点": "确认继续所有人工确认节点",
+    "请先选择待回档集群": "请先选择待回档集群",
+    "请先添加待回档集群": "请先添加待回档集群",
     "任务": "任务",
     "Redis版本": "Redis版本",
     "当前分片数": "当前分片数",
