+++ dbm-ui/frontend/src/views/db-manage/mysql/rollback/pages/page1/Index.vue
@@ -25,10 +25,10 @@
     type="card"
     @change="handleChange">
     <BkRadioButton
-      v-for="(item, index) in rollbackInfos"
-      :key="index"
-      :label="item.value">
-      {{ item.label }}
+      v-for="(value, key) in rollbackTypeLabel"
+      :key="key"
+      :label="key">
+      {{ value }}
     </BkRadioButton>
   </BkRadioGroup>
   <RenderData
@@ -36,6 +36,7 @@
     :data="tableData"
     :rollback-cluster-type="rollbackClusterType" />
 </template>
+
 <script setup lang="ts">
   import { useI18n } from 'vue-i18n';
 
@@ -47,8 +48,8 @@
 
   import TimeZonePicker from '@components/time-zone-picker/index.vue';
 
-  import { rollbackInfos } from './components/common/const';
-  import RenderData, { createRowData, type IDataRow } from './components/render-data/Index.vue';
+  import { createRowData, type IDataRow } from './components/render-row/Index.vue';
+  import RenderData from './components/RenderData.vue';
 
   const { t } = useI18n();
 
@@ -57,14 +58,20 @@
     type: TicketTypes.MYSQL_ROLLBACK_CLUSTER,
     onSuccess(cloneData) {
       rollbackClusterType.value = cloneData.rollback_cluster_type;
-      tableData.value = cloneData.tableDataList as IDataRow[];
+      tableData.value = cloneData.tableDataList;
       window.changeConfirm = true;
     },
   });
 
+  const rollbackTypeLabel = {
+    [RollbackClusterTypes.BUILD_INTO_NEW_CLUSTER]: t('构造到新集群'),
+    [RollbackClusterTypes.BUILD_INTO_EXIST_CLUSTER]: t('构造到已有集群'),
+    [RollbackClusterTypes.BUILD_INTO_METACLUSTER]: t('构造到原集群'),
+  };
+
   const renderDataRef = ref<InstanceType<typeof RenderData>>();
   const rollbackClusterType = ref<RollbackClusterTypes>(RollbackClusterTypes.BUILD_INTO_NEW_CLUSTER);
-  const tableData = shallowRef<Array<IDataRow>>([createRowData({})]);
+  const tableData = shallowRef<IDataRow[]>([createRowData({})]);
 
   const handleChange = () => {
     renderDataRef.value!.reset();
