+++ dbm-ui/frontend/src/views/db-manage/mysql/rollback/pages/page1/components/render-row/components/RenderBackup.vue
@@ -20,13 +20,25 @@
     :rules="rules"
     @change="(value) => handleChange(value as BackupSources)" />
 </template>
-<script setup lang="ts">
-  import { ref } from 'vue';
-  import { useI18n } from 'vue-i18n';
 
-  import TableEditSelect from '@components/render-table/columns/select/index.vue';
+<script lang="ts">
+  import { t } from '@locales/index';
 
-  import { BackupSources, selectList } from './const';
+  export enum BackupSources {
+    REMOTE = 'remote',
+    LOCAL = 'local',
+  }
+
+  export const backupSourceList = [
+    {
+      value: BackupSources.REMOTE,
+      label: t('远程备份'),
+    },
+    {
+      value: BackupSources.LOCAL,
+      label: t('本地备份'),
+    },
+  ];
 
   interface Props {
     modelValue: BackupSources;
@@ -43,21 +55,31 @@
   interface Exposes {
     getValue: () => Promise<Record<string, string>>;
   }
+</script>
+<script setup lang="ts">
+  import { ref } from 'vue';
+
+  import TableEditSelect from '@components/render-table/columns/select/index.vue';
 
   const props = defineProps<Props>();
+
   const emits = defineEmits<Emits>();
 
-  const { t } = useI18n();
   const rules = [
     {
       validator: (value: string) => Boolean(value),
       message: t('备份源不能为空'),
     },
   ];
-  const targetList = props.list || selectList.backupSource;
+  const targetList = props.list || backupSourceList;
 
   const editSelectRef = ref<InstanceType<typeof TableEditSelect>>();
   const localValue = ref(BackupSources.REMOTE);
+
+  const handleChange = (value: BackupSources) => {
+    localValue.value = value;
+    emits('change', localValue.value);
+  };
 
   watch(
     () => props.modelValue,
@@ -68,11 +90,6 @@
       immediate: true,
     },
   );
-
-  const handleChange = (value: BackupSources) => {
-    localValue.value = value;
-    emits('change', localValue.value);
-  };
 
   defineExpose<Exposes>({
     getValue() {
