+++ dbm-ui/frontend/src/views/db-manage/mysql/rollback/pages/page1/components/render-row/components/RenderCluster.vue
@@ -18,15 +18,19 @@
     :placeholder="placeholder || t('请输入集群域名或从表头批量选择')"
     :rules="rules" />
 </template>
+
 <script lang="ts">
   const clusterIdMemo: { [key: string]: Record<string, boolean> } = {};
+
   interface Props {
     modelValue?: IDataRow['clusterData'];
     placeholder?: string;
   }
+
   interface Emits {
-    (e: 'change', data: Props['modelValue']): void;
+    (e: 'change', data: IDataRow['clusterData']): void;
   }
+
   interface Exposes {
     getValue: () => Promise<{
       cluster_id: number;
@@ -39,30 +43,30 @@
 
   import { queryClusters } from '@services/source/mysqlCluster';
 
-  import { useGlobalBizs } from '@stores';
-
   import TableEditInput from '@components/render-table/columns/input/index.vue';
 
   import { random } from '@utils';
 
-  import type { IDataRow } from '../render-data/Index.vue';
+  import type { IDataRow } from '../Index.vue';
 
   const props = withDefaults(defineProps<Props>(), {
-    modelValue: undefined,
+    modelValue: () => ({
+      id: 0,
+      domain: '',
+    }),
     placeholder: '',
   });
+
   const emits = defineEmits<Emits>();
 
-  const instanceKey = `render_cluster_${random()}`;
-  clusterIdMemo[instanceKey] = {};
-
-  const { currentBizId } = useGlobalBizs();
   const { t } = useI18n();
 
   const editRef = ref<InstanceType<typeof TableEditInput>>();
-
   const localClusterId = ref(0);
   const localDomain = ref('');
+
+  const instanceKey = `render_cluster_${random()}`;
+  clusterIdMemo[instanceKey] = {};
 
   const rules = [
     {
@@ -82,7 +86,7 @@
               immute_domain: domain,
             },
           ],
-          bk_biz_id: currentBizId,
+          bk_biz_id: window.PROJECT_CONFIG.BIZ_ID,
         }).then((data) => {
           if (data.length > 0) {
             const {
