+++ dbm-ui/frontend/src/views/db-manage/mysql/rollback/pages/page1/components/render-row/components/RenderHostInputSelect.vue
@@ -22,13 +22,12 @@
       @click-seletor="handleShowIpSelector" />
   </div>
   <IpSelector
-    v-if="clusterData"
     v-model:show-dialog="isShowIpSelector"
     :biz-id="currentBizId"
     button-text=""
     :cloud-info="{
-      id: clusterData.cloudId,
-      name: clusterData.cloudName,
+      id: clusterData!.cloudId,
+      name: clusterData!.cloudName,
     }"
     :data="localHostList"
     :os-types="[OSTypes.Linux]"
@@ -37,9 +36,11 @@
     :single-host-select="single"
     @change="handleHostChange" />
 </template>
+
 <script setup lang="ts">
   import { useI18n } from 'vue-i18n';
 
+  import type { RollbackHost } from '@services/model/ticket/details/mysql';
   import { checkHost } from '@services/source/ipchooser';
   import type { HostInfo } from '@services/types';
 
@@ -52,30 +53,24 @@
 
   import TableSeletorInput from '@views/db-manage/common/TableSeletorInput.vue';
 
-  import type { IDataRow } from '../render-data/Index.vue';
+  import { messageWarn } from '@utils';
 
-  export interface HostDataItem {
-    bk_host_id: number;
-    ip: string;
-    bk_cloud_id: number;
-    bk_biz_id: number;
-  }
+  import type { IDataRow } from '../Index.vue';
 
   interface Props {
     clusterData: IDataRow['clusterData'];
     single?: boolean;
-    hostData: HostDataItem[];
+    hostData: RollbackHost[];
   }
 
   interface Exposes {
     getValue: () => Promise<{
-      hosts: HostDataItem[];
+      hosts: RollbackHost[];
     }>;
   }
 
   const props = withDefaults(defineProps<Props>(), {
     single: false,
-    hostData: undefined,
   });
 
   const { t } = useI18n();
@@ -83,7 +78,7 @@
 
   const rules = [
     {
-      validator: (value: string) => Boolean(value),
+      validator: (value: string) => !!value,
       message: t('IP 不能为空'),
     },
     {
@@ -92,12 +87,12 @@
     },
     {
       validator: async (value: string) => {
-        const list = value.split(batchSplitRegex);
+        const ips = value.split(batchSplitRegex);
         return await checkHost({
-          ip_list: list,
+          ip_list: ips,
         }).then((data) => {
-          if (data.length === list.length) {
-            allocatData(data);
+          if (data.length === ips.length) {
+            localHostList.value = data;
             return true;
           }
           return false;
@@ -107,39 +102,40 @@
     },
   ];
 
-  const editRef = ref();
+  const editRef = ref<InstanceType<typeof TableSeletorInput>>();
   const isShowIpSelector = ref(false);
-  const localValue = ref();
+  const localValue = ref('');
   const localHostList = shallowRef<HostInfo[]>([]);
 
   const placeholder = computed(() => (props.single ? t('请输入或选择 (1台)') : t('请输入或选择')));
 
   const handleShowIpSelector = () => {
+    if (!props.clusterData?.id) {
+      messageWarn(t('请先选择待回档集群'));
+      return;
+    }
     isShowIpSelector.value = true;
   };
 
   // 批量选择
-  const handleHostChange = (hostList: HostInfo[]) => {
-    allocatData(hostList);
-    window.changeConfirm = true;
-    setTimeout(() => {
-      editRef.value.getValue();
-    });
-  };
-
-  const allocatData = (data: HostInfo[]) => {
-    localValue.value = data.map((item) => item?.ip).join(',');
+  const handleHostChange = (data: HostInfo[]) => {
+    localValue.value = data.map((item) => item.ip).join(',');
     localHostList.value = data;
+    window.changeConfirm = true;
   };
 
   watch(
     () => props.hostData,
     (data) => {
-      const fristHost = data?.[0];
-      if (!fristHost?.ip) {
-        return;
+      if (data.length > 0) {
+        const ips = data.map((item) => item.ip).join(',');
+        if (ips) {
+          localValue.value = ips;
+          setTimeout(() => {
+            editRef.value!.getValue();
+          });
+        }
       }
-      allocatData(data as unknown as HostInfo[]);
     },
     {
       immediate: true,
@@ -152,8 +148,8 @@
         hostList.map((item) => ({
           bk_host_id: item.host_id,
           ip: item.ip,
-          bk_cloud_id: item.cloud_area.id,
-          bk_biz_id: currentBizId,
+          bk_cloud_id: item.cloud_area?.id,
+          bk_biz_id: item.biz?.id,
         }));
 
       return Promise.resolve({
@@ -162,6 +158,7 @@
     },
   });
 </script>
+
 <style lang="less" scoped>
   .render-host-box {
     position: relative;
