+++ dbm-ui/frontend/src/views/tickets/common/components/demand-factory/mysql/rollback/Index.vue
@@ -57,9 +57,7 @@
     },
   };
 
-  const renderData = computed(
-    () => rollbackInfo[props.ticketDetails.details.rollback_cluster_type as RollbackClusterTypes],
-  );
+  const renderData = computed(() => rollbackInfo[props.ticketDetails.details.rollback_cluster_type]);
 </script>
 
 <style lang="less" scoped>
