+++ dbm-ui/frontend/src/views/tickets/common/components/demand-factory/spider/rollback/components/RollbackNewCluster.vue
@@ -21,16 +21,15 @@
 <script setup lang="tsx">
   import { useI18n } from 'vue-i18n';
 
-  import type { MySQLRollbackDetails } from '@services/model/ticket/details/mysql';
+  import type { SpiderRollbackDetails } from '@services/model/ticket/details/spider';
   import TicketModel from '@services/model/ticket/ticket';
 
-  import { type BackupSources, selectList } from '@views/db-manage/mysql/rollback/pages/page1/components/common/const';
-  import type { IDataRow } from '@views/db-manage/tendb-cluster/rollback/pages/page1/components/render-data/Index.vue';
+  import { backupSourceList, type BackupSources } from '@views/db-manage/mysql/rollback/pages/page1/components/render-row/components/RenderBackup.vue';
 
   import { utcDisplayTime } from '@utils';
 
   interface Props {
-    ticketDetails: TicketModel<MySQLRollbackDetails>
+    ticketDetails: TicketModel<SpiderRollbackDetails>
   }
 
   const props = defineProps<Props>();
@@ -54,25 +53,25 @@
       label: t('存储层主机'),
       field: 'remote_hosts',
       width: 120,
-      render: ({ data }: { data: MySQLRollbackDetails['infos'][0] }) => <span>{(data.rollback_host as IDataRow['rollbackHost']).remote_hosts?.map(item => item.ip) || '--'}</span>,
+      render: ({ data }: { data: SpiderRollbackDetails['infos'][0] }) => <span>{data.rollback_host.remote_hosts?.map(item => item.ip) || '--'}</span>,
     },
     {
       label: t('接入层主机'),
       field: 'spider_host',
       width: 120,
-      render: ({ data }: { data: MySQLRollbackDetails['infos'][0] }) => <span>{(data.rollback_host as IDataRow['rollbackHost']).spider_host?.ip || '--'}</span>,
+      render: ({ data }: { data: SpiderRollbackDetails['infos'][0] }) => <span>{data.rollback_host.spider_host?.ip || '--'}</span>,
     },
     {
       label: t('备份源'),
       field: 'backup_source',
       width: 100,
-      render: ({ cell }: { cell: BackupSources }) => <span>{ selectList.backupSource.find(item=>item.value === cell)?.label || '--' }</span>,
+      render: ({ cell }: { cell: BackupSources }) => <span>{ backupSourceList.find(item=>item.value === cell)?.label || '--' }</span>,
     },
     {
       label: t('回档类型'),
       field: '',
       width: 280,
-      render: ({ data }: { data: MySQLRollbackDetails['infos'][0] }) =>  {
+      render: ({ data }: { data: SpiderRollbackDetails['infos'][0] }) =>  {
         if (data.rollback_time) {
           return <span>{ t('回档到指定时间') } - { utcDisplayTime(data.rollback_time) }</span>;
         }
